feat: print backtrace when resulting errors (#215)
It is said in the [`anyhow::Error`
docs](https://docs.rs/anyhow/latest/anyhow/struct.Error.html#display-representations)
> The Debug format “{:?}” includes your backtrace if one was captured.
Note that this is the representation you get by default if you return an
error from fn main instead of printing it explicitly yourself.

Usually the `Error` is handle with `.unwrap` or other functions which
finally call `panic`. And the panic handler in rust std will give a
friendly tip if the backtrace is disabled. However, `cargo-shear`
doesn't call panic but customize its exit code instead, so that's the
reason why I also add this tip. Feel free to remove it if you think this
is a bad design :)